+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Minor changes
 * Changed default output folder to `PhishBench Output` 
+* Moved log and summary to output folder
 
 # v1.3.0