+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Minor changes
 * Changed default output folder to `PhishBench Output` 
 * Moved log and summary to output folder
+* Updated user interaction
 
 # v1.3.0